pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-family"
version = "0.1.0"
description = "Google ADK multi-agent system with A2A protocol, Master Agent orchestrating CalendarAgent and TaskAgent"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "Agent Family" }]

dependencies = [
    "google-adk>=1.28.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "rich>=13.7.0",
    # Real Google API tools
    "fastmcp>=2.3.0",
    "google-api-python-client>=2.120.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "cryptography>=42.0.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "anyio[trio]>=4.3.0",
    "faker>=25.0.0",
    "responses>=0.25.0",
    "freezegun>=1.4.0",
]

[project.scripts]
agent-family = "agent_family.runner:cli_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_family*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
    "unit: fast, dependency-free unit tests",
    "integration: tests that require API keys",
    "routing: Master Agent intent-routing tests",
]

[tool.ruff]
line-length = 100
target-version = "py312"
